move gridlayouter to its own file
